requirements.txt
# DarwinLM - Evolutionary Prisoner's Dilemma Tournament
# Core dependencies
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Async utilities
asyncio-throttle>=1.0.2
aiohttp>=3.9.0

# Visualization
streamlit>=1.30.0
plotly>=5.18.0

# CLI and logging
rich>=13.0.0
typer>=0.9.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
